Fleshed out base functions
Used Clang.jl to wrap librcl, librmw, and librcutils. Then, updated
the the library to use those wrappers, and now all the tests pass
for init(), ok(), shutdown(), and instance_id().

Note: the autogen wrapping code only works on Julia 0.6 for now.
But only the developer of this package (me) has to ever run it, so
I don't care too much about fixing it for 0.7.